fix : UserMode, AdminMode에 따른 입/출력 클래스 분리
