+++ VendingMachine/Views/InputView/InputView.swift
@@ -24,13 +24,7 @@ class InputView {
         }
         return (numberArray[0], numberArray[1])
     }
-    static func readAdminInstruction() throws -> Int {
-        guard let instruction = readLine() else {
-            throw VendingMachineError.invalidMenuSelectNumberError
-        }
-        let number = try Validation.convertStringToNumber(instruction)
-        return number
-    }
+   
     static func readModeSelection() throws -> Int{
         guard let modeSelection = readLine() else{
             throw VendingMachineError.invalidMenuSelectNumberError
